requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
aio-pika==9.4.1
structlog==23.2.0
httpx==0.25.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
flake8==6.1.0
mypy==1.6.1
testcontainers==4.4.1
aiofiles==23.2.1